Split Sudoku example
This commit splits the Sudoku example into a simple
examples/sudoku.cpp that contains the straightforward simple Sudoku
model, and a examples/sudoku-advanced.cpp that contains the more
advanced model that can mix integer and set constraints and has all
the different branching variants. The main motivation is pedagogial,
since Sudoku is one of the first examples to show and look at in a
system.

The instances are extracted to examples/sudoku-instances.hpp to be
useful from both examples.
