+++ Makefile.in
@@ -865,7 +865,7 @@ INTEXAMPLEHDR0 =						\
 INTEXAMPLESRC0 =						\
 	alpha bacp bibd donald efpa eq20 golomb-ruler		\
 	graph-color grocery ind-set magic-sequence magic-square	\
-	money ortho-latin partition photo queens sudoku kakuro  \
+	money ortho-latin partition photo queens sudoku sudoku-advanced kakuro  \
 	nonogram pentominoes crowded-chess black-hole		\
 	minesweeper domino steel-mill sports-league		\
 	all-interval langford-number warehouses	radiotherapy    \
@@ -1288,7 +1288,7 @@ makecompletedmessage:
 	@echo make install
 	@echo
 
-# ugly hack by Grégoire Dooms to call a target on a contribs after evaluating all the variables in this Makefile.
+# ugly hack by Grï¿½goire Dooms to call a target on a contribs after evaluating all the variables in this Makefile.
 # used as make contribs:cpgraph:doc or make contribs:cpgraph:Makefile.in
 .PHONY: contribs\:%
 contribs\:%:
